+++ src/menu_button.py
@@ -2,7 +2,7 @@ from telegram import BotCommand
 from telegram.ext import CallbackContext
 
 from constants import commands
-from texts import BTN_MENU2, BTN_START
+from texts.buttons import BTN_MENU2, BTN_START
 
 COMMANDS_UNAUTHORIZED = (BotCommand(commands.START, BTN_START),)
 COMMANDS = (
