+++ src/texts/buttons.py
@@ -1,0 +1,35 @@
+BTN_YES = "Да"
+BTN_NO = "Нет"
+BTN_MONTH_STAT = "Статистика за месяц"
+BTN_IN_PROGRESS = "В работе"
+BTN_OVERDUE = "🔥 Cроки горят"
+BTN_RULES = "Правила сервиса"
+
+
+# conversation/authorization.py
+BTN_CONSULTATION = "Получить онлайн-консультацию"
+BTN_SUPPORT = "Поддержать проект"
+
+
+# conversation/menu.py
+BTN_TIMEZONE = "Настроить часовой пояс (сейчас {user_tz})"
+BTN_MENU = "Меню"
+
+# conversation/timezone.py
+BTN_GEOLOCATION = "Отправить свою геолокацию для автоматической настройки часового пояса"
+BTN_TIMEZONE_DEFAULT = "Таймзона по умолчанию UTC+03:00 (Москва)."
+BTN_TIMEZONE_BY_LOCATION_OR_MANUAL = "Таймзона по локации или вручную."
+
+
+# services/repeat_message.py
+BTN_REMIND_IN_HOUR = "🕑 Напомнить через час"
+
+
+# jobs.py
+BTN_BILL_SENT = "✅ Уже отправил(а)"
+BTN_BILL_SOON = "🕑 Скоро отправлю"
+
+
+# menu_button.py
+BTN_START = "Начать работу с ботом."
+BTN_MENU2 = "Показать список возможных запросов."
