+++ src/texts/common.py
@@ -1,0 +1,9 @@
+__JUST_ASK = '"Просто спросить"'
+__OPEN_TRELLO = "[Открыть Trello]({trello_url})"
+
+PLURAL_DAY = "день", "дня", "дней"
+PLURAL_HOUR = "час", "часа", "часов"
+PLURAL_CONSULTATION_NOT_SINGLE = "заявки", "заявок"
+PLURAL_CONSULTATION = "заявка", *PLURAL_CONSULTATION_NOT_SINGLE
+
+DATA_NOT_AVAILABLE = "Данные недоступны!"
